Added payment form
Payment form was added with email, phone number, card number and exp date validators.

Form also includes a function that makes sure that each input is filled in, otherwise the submission fails and an asterisk gets added along with an error message.

Phone number validator still needs improvement to only allow numbers to be entered and only 10 digits not more